renames SinglePost.js to EditListing.js
